Moved construction of graph to its own component. This is a preparation for adding features to the engine.
