+++ engine-core/src/main/scala/org/scalarules/derivations/DerivationGraph.scala
@@ -1,7 +1,10 @@
-package org.scalarules.engine
+package org.scalarules.derivations
+
+import org.scalarules.engine.Derivation
 
 trait DerivationGraph {
 
 }
 
 case class Node(derivation: Derivation, children: List[Node]) extends DerivationGraph
+
