+++ engine-core/src/main/scala/org/scalarules/engine/evaluations.scala
@@ -1,5 +1,7 @@
 package org.scalarules.engine
 
+import org.scalarules.facts.Fact
+
 trait Evaluation[+A] {
   def apply(c: Context): Option[A]
