+++ engine-core/src/main/scala/org/scalarules/facts/facts.scala
@@ -1,4 +1,6 @@
-package org.scalarules.engine
+package org.scalarules.facts
+
+import org.scalarules.engine.{ErrorEvaluation, Evaluation, ListFactEvaluation, SingularFactEvaluation}
 
 import scala.language.existentials
 
@@ -13,19 +15,19 @@ trait Fact[+A] {
   override def toString: String = name
 }
 
-private[engine] case object OriginFact extends Fact[Nothing] {
-  def name: String = "___meta___OriginFact___meta___"
-  def description: String = "Meta-fact used in graph construction"
-
-  def toEval: Evaluation[Nothing] = new ErrorEvaluation("The OriginFact is a meta-fact used in graph construction to indicate top-level constant evaluations")
-
-  def valueType: String = "Nothing"
-}
-
 case class SingularFact[+A](name: String, description: String = "", valueType: String = "") extends Fact[A] {
   def toEval: Evaluation[A] = new SingularFactEvaluation(this)
 }
 
 case class ListFact[+A](name: String, description: String = "", valueType: String = "") extends Fact[List[A]] {
   def toEval: Evaluation[List[A]] = new ListFactEvaluation[A](this)
+}
+
+case object OriginFact extends Fact[Nothing] {
+  def name: String = "___meta___OriginFact___meta___"
+  def description: String = "Meta-fact used in graph construction"
+
+  def toEval: Evaluation[Nothing] = new ErrorEvaluation("The OriginFact is a meta-fact used in graph construction to indicate top-level constant evaluations")
+
+  def valueType: String = "Nothing"
 }
