+++ engine-test-utils/src/main/scala/org/scalarules/testutils/nl/HeuristicTester.scala
@@ -1,7 +1,8 @@
 package org.scalarules.testutils.nl
 
 import org.scalarules.dsl.nl.grammar.Berekening
-import org.scalarules.engine.{Context, Fact, FactEngine}
+import org.scalarules.engine.{Context, FactEngine}
+import org.scalarules.facts.Fact
 import org.scalarules.services._
 import org.scalarules.utils.PrettyPrinter
