+++ engine/src/test/scala/org/scalarules/utils/InternalTestUtils.scala
@@ -1,6 +1,7 @@
 package org.scalarules.utils
 
-import org.scalarules.engine.{Derivation, Fact, FactEngine, Step, _}
+import org.scalarules.engine.{Derivation, FactEngine, Step, _}
+import org.scalarules.facts.Fact
 import org.scalarules.services.{AnalyzedScenario, HeuristicService}
 
 object InternalTestUtils {
